Updated with 4 scenes now working.

+++ src/FftTrigger.h
@@ -27,8 +27,8 @@ class FftTrigger
         float value ;
         ofColor color ;
     
-        int lowBand;
-        int highBand;
+        float lowBand;
+        float highBand;
         float height;
         string name;
